comments in 1up% file, fidgetting in 1shm
added notes on the goals of this TAS and the berry combo rules to the full game file. Also some notes on when collection can theoretically begin on the first map. also started working on removing the first berry collection, which is way more annoying than I want it to be.